+++ packages/cubejs-duckdb-driver/src/DuckDBDriver.ts
@@ -2,12 +2,13 @@ import {
   BaseDriver,
   DriverInterface,
   StreamOptions,
-  QueryOptions, StreamTableData,
+  QueryOptions,
+  StreamTableData,
+  GenericDataBaseType,
 } from '@cubejs-backend/base-driver';
 import { getEnv } from '@cubejs-backend/shared';
 import { promisify } from 'util';
 import * as stream from 'stream';
-// eslint-disable-next-line import/no-extraneous-dependencies
 import { Connection, Database } from 'duckdb';
 
 import { DuckDBQuery } from './DuckDBQuery';
@@ -20,8 +21,15 @@ export type DuckDBDriverConfiguration = {
 };
 
 type InitPromise = {
-  connection: Connection,
+  defaultConnection: Connection,
   db: Database;
+};
+
+const DuckDBToGenericType: Record<string, GenericDataBaseType> = {
+  // DATE_TRUNC returns DATE, but Cube Store still doesn't support DATE type
+  // DuckDB's driver transform date/timestamp to Date object, but HydrationStream converts any Date object to ISO timestamp
+  // That's why It's safe to use timestamp here
+  date: 'timestamp',
 };
 
 export class DuckDBDriver extends BaseDriver implements DriverInterface {
@@ -37,14 +45,24 @@ export class DuckDBDriver extends BaseDriver implements DriverInterface {
     this.schema = this.config.schema || getEnv('duckdbSchema', this.config);
   }
 
+  public toGenericType(columnType: string): GenericDataBaseType {
+    if (columnType.toLowerCase() in DuckDBToGenericType) {
+      return DuckDBToGenericType[columnType.toLowerCase()];
+    }
+
+    return super.toGenericType(columnType.toLowerCase());
+  }
+
   protected async init(): Promise<InitPromise> {
     const token = getEnv('duckdbMotherDuckToken', this.config);
     
     const db = new Database(token ? `md:?motherduck_token=${token}` : ':memory:');
-    const connection = db.connect();
-    
+    // Under the hood all methods of Database uses internal default connection, but there is no way to expose it
+    const defaultConnection = db.connect();
+    const execAsync: (sql: string, ...params: any[]) => Promise<void> = promisify(defaultConnection.exec).bind(defaultConnection) as any;
+
     try {
-      await this.handleQuery(connection, 'INSTALL httpfs', []);
+      await execAsync('INSTALL httpfs');
     } catch (e) {
       if (this.logger) {
         console.error('DuckDB - error on httpfs installation', {
@@ -57,7 +75,7 @@ export class DuckDBDriver extends BaseDriver implements DriverInterface {
     }
 
     try {
-      await this.handleQuery(connection, 'LOAD httpfs', []);
+      await execAsync('LOAD httpfs');
     } catch (e) {
       if (this.logger) {
         console.error('DuckDB - error on loading httpfs', {
@@ -99,7 +117,7 @@ export class DuckDBDriver extends BaseDriver implements DriverInterface {
     for (const { key, value } of configuration) {
       if (value) {
         try {
-          await this.handleQuery(connection, `SET ${key}='${value}'`, []);
+          await execAsync(`SET ${key}='${value}'`);
         } catch (e) {
           if (this.logger) {
             console.error(`DuckDB - error on configuration, key: ${key}`, {
@@ -112,7 +130,7 @@ export class DuckDBDriver extends BaseDriver implements DriverInterface {
 
     if (this.config.initSql) {
       try {
-        await this.handleQuery(connection, this.config.initSql, []);
+        await execAsync(this.config.initSql);
       } catch (e) {
         if (this.logger) {
           console.error('DuckDB - error on init sql (skipping)', {
@@ -123,7 +141,7 @@ export class DuckDBDriver extends BaseDriver implements DriverInterface {
     }
     
     return {
-      connection,
+      defaultConnection,
       db
     };
   }
@@ -148,14 +166,13 @@ export class DuckDBDriver extends BaseDriver implements DriverInterface {
     return super.getSchemasQuery();
   }
 
-  protected async getConnection(): Promise<Connection> {
+  protected async getInitiatedState(): Promise<InitPromise> {
     if (!this.initPromise) {
       this.initPromise = this.init();
     }
 
     try {
-      const { connection } = await this.initPromise;
-      return connection;
+      return await this.initPromise;
     } catch (e) {
       this.initPromise = null;
 
@@ -167,15 +184,11 @@ export class DuckDBDriver extends BaseDriver implements DriverInterface {
     return DuckDBQuery;
   }
 
-  protected handleQuery<R>(connection: Connection, query: string, values: unknown[] = [], _options?: QueryOptions): Promise<R[]> {
-    const executeQuery: (sql: string, ...args: any[]) => Promise<R[]> = promisify(connection.all).bind(connection) as any;
-
-    return executeQuery(query, ...values);
-  }
-
   public async query<R = unknown>(query: string, values: unknown[] = [], _options?: QueryOptions): Promise<R[]> {
-    const result = await this.handleQuery<R>(await this.getConnection(), query, values, _options);
+    const { defaultConnection } = await this.getInitiatedState();
+    const fetchAsync: (sql: string, ...params: any[]) => Promise<R[]> = promisify(defaultConnection.all).bind(defaultConnection) as any;
 
+    const result = await fetchAsync(query, ...values);
     return result.map((item) => {
       transformRow(item);
 
@@ -188,14 +201,29 @@ export class DuckDBDriver extends BaseDriver implements DriverInterface {
     values: unknown[],
     { highWaterMark }: StreamOptions
   ): Promise<StreamTableData> {
-    const connection = await this.getConnection();
+    const { db } = await this.getInitiatedState();
 
-    const asyncIterator = connection.stream(query, ...(values || []));
-    const rowStream = stream.Readable.from(asyncIterator, { highWaterMark }).pipe(new HydrationStream());
+    // new connection, because stream can break with
+    // Attempting to execute an unsuccessful or closed pending query result
+    // PreAggregation queue has a concurrency limit, it's why pool is not needed here
+    const connection = db.connect();
+    const closeAsync = promisify(connection.close).bind(connection);
 
-    return {
-      rowStream,
-    };
+    try {
+      const asyncIterator = connection.stream(query, ...(values || []));
+      const rowStream = stream.Readable.from(asyncIterator, { highWaterMark }).pipe(new HydrationStream());
+
+      return {
+        rowStream,
+        release: async () => {
+          await closeAsync();
+        }
+      };
+    } catch (e) {
+      await closeAsync();
+
+      throw e;
+    }
   }
 
   public async testConnection(): Promise<void> {