+++ rust/cubesql/cubesql/src/compile/rewrite/rules/wrapper/mod.rs
@@ -7,14 +7,19 @@ mod cast;
 mod column;
 mod cube_scan_wrapper;
 mod extract;
+mod in_list_expr;
 mod is_null_expr;
 mod limit;
 mod literal;
+mod negative_expr;
+mod not_expr;
 mod order;
 mod projection;
 mod scalar_function;
 mod sort_expr;
 mod udf_function;
+mod window;
+mod window_function;
 mod wrapper_pull_up;
 
 use crate::compile::{
@@ -44,7 +49,9 @@ impl RewriteRules for WrapperRules {
         self.projection_rules(&mut rules);
         self.limit_rules(&mut rules);
         self.order_rules(&mut rules);
+        self.window_rules(&mut rules);
         self.aggregate_function_rules(&mut rules);
+        self.window_function_rules(&mut rules);
         self.scalar_function_rules(&mut rules);
         self.udf_function_rules(&mut rules);
         self.extract_rules(&mut rules);
@@ -56,6 +63,9 @@ impl RewriteRules for WrapperRules {
         self.cast_rules(&mut rules);
         self.column_rules(&mut rules);
         self.literal_rules(&mut rules);
+        self.in_list_expr_rules(&mut rules);
+        self.negative_expr_rules(&mut rules);
+        self.not_expr_rules(&mut rules);
 
         rules
     }