+++ rust/cubesql/cubesql/src/compile/rewrite/rules/wrapper/projection.rs
@@ -4,9 +4,9 @@ use crate::{
         transforming_rewrite, wrapped_select, wrapped_select_aggr_expr_empty_tail,
         wrapped_select_filter_expr_empty_tail, wrapped_select_group_expr_empty_tail,
         wrapped_select_having_expr_empty_tail, wrapped_select_joins_empty_tail,
-        wrapped_select_order_expr_empty_tail, wrapper_pullup_replacer, wrapper_pushdown_replacer,
-        LogicalPlanLanguage, ProjectionAlias, WrappedSelectAlias, WrappedSelectUngrouped,
-        WrapperPullupReplacerUngrouped,
+        wrapped_select_order_expr_empty_tail, wrapped_select_window_expr_empty_tail,
+        wrapper_pullup_replacer, wrapper_pushdown_replacer, LogicalPlanLanguage, ProjectionAlias,
+        WrappedSelectAlias, WrappedSelectUngrouped, WrapperPullupReplacerUngrouped,
     },
     var, var_iter,
 };
@@ -50,6 +50,12 @@ impl WrapperRules {
                     ),
                     wrapper_pullup_replacer(
                         wrapped_select_aggr_expr_empty_tail(),
+                        "?alias_to_cube",
+                        "?ungrouped",
+                        "?cube_members",
+                    ),
+                    wrapper_pullup_replacer(
+                        wrapped_select_window_expr_empty_tail(),
                         "?alias_to_cube",
                         "?ungrouped",
                         "?cube_members",